rewritten footer
